move country list to field-telephone
